Refactored code into smaller files
